Updating word checks
Trying to update the script such that it actually checks each word in the script so that we can exclude images that are just pictures of people or selfies and other bad things we don't want. Currently I think it's comparing a single word string to a string of a bunch of words which is not ideal. It should check every word in the string to make sure there are no matching components. I think I also need to implement some kind of set or list comprehension here but I'll look into it here in a bit.