Update ArtGrabber-Pexels-Pi.py
added a ".lower()" function to the end of the photo description text so that we can match the lower case script strings of the bad words list to the lower case strings of the photo alt text.

This way we can ensure that a word like "Test" gets filtered out the same way that "test" would. If "test" is in the bad words list but "Test" isn't.