Making the previous 2 scripts redundant
Basically in this case, instead of logging photos to the spreadsheet to use later, it just calls Pexels API, searches for a photo to use, and then posts it to FB directly. This singular script will replace the previous 2 scripts.

I will still keep the scripts around in case of emergencies so that we still have a backlog of photos to use.