Fixing a bug where fetchall was throwing an exception
Basically the issue was that if the post ID or Image Hash wasn't in the database, the fetchall would return nothing, which is good, but I was trying to index fetchall when it was an empty list.

Then I realized I really don't care what it returns from the fetchall as long as it returns something or not, So I got rid of the indexing basically. EZ fix.