Using re.sub to strip spec chars out of OCR text
Trying to remove instances where a word could be like "F!!!AT" and still not get flagged. This might not come up very often but who knows. I imagine it would be more likely if a bad word is hyphened to another word and not get caught. Idk but we'll see what happens with it.