FIXED A HILARIOUS BUG (LMAO)
I accidentally was searching the bad words as my key terms instead of the photo search terms that it was supposed to use.

I literally had to go in and change one line of code to fix it but it was hilarious nonetheless. haha

Note: I found this by debugging the code. It's very important to know what your variable values are and how to use your debugger in your IDE. Thanks JetBrains, good job with PyCharm! Really saved me a ton of time here!