add type hinting and added 2 more  things
Added an if statement to control for if the description is empty to get around the bad words check. This is more of a bandaid fix that I'll come back to later.

Also, added a fix for if there is an API error of FB servers are down, now it won't try to post endlessly forever. So now it will only try to post 5 times and then give up until the next hour.