Updating the photo description issues
Honestly the code looks a little weird ngl but it works. I did update the pexels api wrapper code as well to do this, but eh. Need to update pypi, reinstall the package / upgrade the package and all that jazz, just kind of a pain so I'll do it later.