adding some basic print debugging messages
Turns out the bot's API key is broken because of some kind of weird permissions error. I think when the API key went invalid and I got a new one I must have done something wrong in the process to not give it proper permissions with that token. So I'll have to go through that process once more. (yay. lol).

Should have this bot fully functioning soon though. But luckily the bot is working perfectly fine.

It's a bit slow though. I found it it only checks about 1 photo per second. I turned off OCR to help with performance but it's still kinda slow. Not sure about what specifically is the slowest processing. I'm gonna have to do a little bit extra printing with the memory usage and timeit data to see which part of the process takes the longest to see if we can prune it out completely or improve its performance.

I imagine with OCR disabled that hashing is probably the slowest part as the rest is basically just string manipulation, but I'm not entirely sure. Something to look into later on eventually.